+++ Cpp/plane_boarding/include/board.hpp
@@ -29,8 +29,8 @@ public:
     Board(unsigned int rows_nr, unsigned int seats_in_row);
 
     // Get info about board
-    unsigned int rows_nr() const { return rows_nr_; }
-    unsigned int seats_nr() const {return seats_nr_; }
+    unsigned int get_rows_nr() const { return rows_nr_; }
+    unsigned int get_seats_nr() const {return seats_nr_; }
     const std::vector<Row>& rows() const {return rows_; }
     const std::vector<std::unique_ptr<PassengerOnBoard>>& aisle() const { return aisle_; }
     const std::vector<std::unique_ptr<PassengerOnBoard>>& outer_queue() const {return outer_queue_; }
@@ -60,6 +60,7 @@ public:
     void generate_tour_report(unsigned int turn, const std::string& report_file_name) const ;
 
 private:
+    // Methods:
     // generating luggage managing time
     unsigned int luggage_time() const;
 
@@ -69,6 +70,7 @@ private:
     void wiki_queue();
     void even_windows_queue();
 
+    // Attributes
     unsigned int rows_nr_;
     unsigned int seats_nr_;
     std::vector<Row> rows_;
@@ -76,19 +78,13 @@ private:
     std::vector<std::unique_ptr<PassengerOnBoard>> outer_queue_;
     std::vector<std::vector<Passenger>> board_plan_;
 
+    // time attributes
     unsigned int fixed_manage_luggage_time_ = 2U;
     unsigned int random_luggage_max_time_ = 4U;
     unsigned int random_luggage_min_time_ = 1U;
-    bool random_luggage_time = false;
+    bool random_luggage_time_ = false;
 };
 
-// Report helpers functions
-size_t digit_nr(unsigned int nr);
-size_t digit_max_index_nr(unsigned int nr);
-size_t get_position_width(const Board& board);
-std::string position_correct_width(unsigned int act_row, unsigned int rows_nr, unsigned int act_seat, unsigned int seats_nr);
-std::string upper_line(const Board& board);
-std::string line_passenger(const Board& board, unsigned int line_nr);
-std::string queue_aisle_line(const Board& board);
+
 
 #endif //PLANE_BOARDING_BOARD_HPP
